Include libShadowDive and kode54/dumb as submodules & integrate with cmake
The goal is to make getting started with OpenOMF easier. All of the
other dependancies are relatively common in operating system package
managers, but libshadowdive and the particular fork of dumb we use are
not. This branch takes care of getting the not-so-common deps working
out of the box. It also helps with reproducible builds, as the version
of libshadowdive is pinned as a submodule.