CSV taglist and generators.
This adds a single location for taglist (a CSV file) and a couple of generators for creating a C list
and sqlite file from it. The point is that we can add the tag descriptions to a single location,
instead of having to modify tons of separate files and databases.