Get shadow grab to mostly work
Remaining issue for shadow grab is releasing the HAR when the shadow
fades and fading the shadow if the grabbed har is hit.

The shadow fading when the har is hit is managed with the 'uz' tag. The
'ua' tag in the grab animation keeps the HAR looping in the struggle
animation.