Rework dialog boxes to use modern components and sizing
Dialog boxes were previously very hardcoded. This PR makes them use the
more modern container widgets and text rendering so they are more
flexible with regard to their contents.

This also improves the menu widget layout behavior and adds support for
centered horizontal menus, which are used for the dialog buttons.

Dialog boxes are now made up of a menu containing a textlabel, for the
dialog text, then a horizontal, centered menu that holds the dialog
buttons.