Make rollback perfectly consistent with local play & improve dedup
When we were deduping inputs, we were not considering the direction the
HAR was facing. This led to a case where if you jumped over the opponent
and kept holding the forward and jump keys, once you landed the HAR, on
rollback, would start jumping the other way. This is because we were
encoding the direction into the input buffer, but not when deduping, so
because the direction of the HAR changed, we need to allow the new input
to come through.

Additionally, the dedup has been corrected so that if the last event the
player has not changed, and the HAR's direction has not changed, there
is no need to store or send any input events. This significantly reduces
the network events sent and thus the rollbacks that need to be done.