Add support for hyper mode and tournament enhancements
Implement the animation string search strategy for
hyper/normal/enhancement animation strings when loading the HAR.

Essentially, in normal 1/2 player mode, if extra string 0 exists and is
not blank or '!', use it. In hyper mode if extra string 1 exists and is
not blank, use it.

For tournament mode, if the user has an enhancement, check index
1+enhancement level, if that exists, use it, otherwise use the highest
populated string in the extra strings table. If the user doesn't have an
enhancement, also honor hyper mode setting.