Implement "walk to HAR" used in several scrap animations
Many of the scrap animations have the winner walk over to where the
loser is lying on the ground and pick them up, or ruin them where they
lay, etc.

Shadow's scrap has the winner walk to the far arena wall and lean
against it.

Both of these are now implemented.

Additionally several bugs with the ez-destruct cheat were fixed, and
going to non-scrap/destruction moves during scrap/destruction are now
properly disallowed.