Store gameplay affecting settings in a game_state struct
This struct avoids direct access to the global settings. This is because
in some cases (tournament mode, netplay, REC playback) we don't want to
use the global settings blindly. For example some options are ignored in
netplay mode (eg player 1/2 power), agreeing on the options for netplay
should be handled differently, and REC files contain their own set of
all the gameplay settings.