Implement input delay for smoother feeling netplay
Allow delay of up to 10 frames on local input. This helps reduce the
jittery feel that can happen with laggy netplay. Additionally allow
delay to be set in local play as well so that people can develop muscle
memory for a given input delay.

Netplay delay is hardcoded to 2 ticks.